ingestion/requirements.txt
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
geoalchemy2==0.14.2
pydantic==2.5.0
pydantic-settings==2.1.0